feat(TypeScript): Allow errors with type unknown to be passed to notify()
TypeScript 4.4 with --strict causes errors in catch() blocks to have
type unknown.

For clients that simply want to pass the error object on to notify()
this can be a nuisance. We already internally treat error objects as
unknown and coerce them into something useful so this just exposes that
functionality to TS clients.

For clients who want to strictly type-check the errors they pass to
notify to ensure that they are passing something that will be sensibly
reported, we expose NotifiableError and allow narrowing the error type
on notify as follows:

```typescript
import { NotifiableError } from '@birchill/bugsnag-zero';

Bugsnag.notify<NotifiableError>(error);
```