fix: handle objects with stringifiers in fetch breadcrumbs plugin
The fetch WebIDL allows a `Request` object or `string` where WebIDL will
automatically call the stringifier on a non-Request object.

This is what makes passing a `URL` to `fetch` work (since the
stringifier for a URL object serializes the object to a URL string).

This patch also makes a few further tweaks to the parameter handling to
more closely reflect the fetch algorithm.

Fixes #132.